Rename rrio package
