+++ sdks/java/io/rrio/build.gradle
@@ -18,7 +18,7 @@
 
 plugins { id 'org.apache.beam.module' }
 applyJavaNature(
-        automaticModuleName: 'org.apache.beam.sdk.io.rrio'
+        automaticModuleName: 'org.apache.beam.sdk.io.requestresponse'
 )
 
 description = "Apache Beam :: SDKS :: Java :: IO :: RequestResponseIO (RRIO)"
