+++ sdks/java/io/rrio/src/main/java/org/apache/beam/io/requestresponse/CacheRead.java
@@ -15,11 +15,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.beam.io.requestresponseio;
+package org.apache.beam.io.requestresponse;
 
 import com.google.auto.value.AutoValue;
 import java.util.Map;
-import org.apache.beam.io.requestresponseio.CacheRead.Result;
+import org.apache.beam.io.requestresponse.CacheRead.Result;
 import org.apache.beam.sdk.Pipeline;
 import org.apache.beam.sdk.transforms.PTransform;
 import org.apache.beam.sdk.values.KV;
