+++ sdks/java/io/rrio/src/test/java/org/apache/beam/io/requestresponse/SetupTeardownTest.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.beam.io.requestresponseio;
+package org.apache.beam.io.requestresponse;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertThrows;
